+++ packages/feature-toggle/src/constants.ts
@@ -14,6 +14,8 @@ export const tokenToggleGuid: ExtensionConfigKeys = {
     'sap.ux.help.testBetaFeatures.showTestGuides': 'fbb03f42-0a86-4fd5-9fc4-8c9b38a4d1a3',
     'sap.ux.help.testBetaFeatures.enableFioriAI': '165a0e31-35ea-4bee-8d47-b8593435a82d',
     'sap.ux.help.testBetaFeatures.enableFioriAIJoule': '165a0e31-35ea-4bee-8d47-b8593435a82e',
+    'sap.ux.help.testBetaFeatures.enableFioriAIVsCode': '165a0e31-35ea-4bee-8d47-b8593435a82f',
+    'sap.ux.help.testBetaFeatures.enableFioriAIAppModeler': '165a0e31-35ea-4bee-8d47-b8593435a82g',
     'sap.ux.applicationModeler.testBetaFeatures.manifestEditor': true,
     'sap.ux.appGenerator.testBetaFeatures.newAnnotationAPI': true
 } as ExtensionConfigKeys;