+++ packages/preview-middleware-client/src/adp/quick-actions/table-quick-action-base.ts
@@ -14,6 +14,7 @@ import { getUi5Version, isLowerThanMinimalUi5Version } from '../../utils/version
 import ObjectPageSection from 'sap/uxap/ObjectPageSection';
 import ObjectPageSubSection from 'sap/uxap/ObjectPageSubSection';
 import ObjectPageLayout from 'sap/uxap/ObjectPageLayout';
+import ManagedObject from 'sap/ui/base/ManagedObject';
 
 const SMART_TABLE_ACTION_ID = 'CTX_COMP_VARIANT_CONTENT';
 const M_TABLE_ACTION_ID = 'CTX_ADD_ELEMENTS_AS_CHILD';
@@ -39,6 +40,10 @@ async function getActionId(table: UI5Element): Promise<string[]> {
 
     return [M_TABLE_ACTION_ID, SETTINGS_ID];
 }
+export type TableQuickActionsOptions = {
+    includeServiceAction?: boolean;
+    areTableRowsRequired?: boolean;
+};
 
 /**
  * Base class for table quick actions.
@@ -85,7 +90,7 @@ export abstract class TableQuickActionDefinitionBase {
         protected readonly controlTypes: string[],
         protected readonly defaultTextKey: string,
         protected readonly context: QuickActionContext,
-        protected includeServiceAction?: boolean
+        protected options: TableQuickActionsOptions = {}
     ) {}
 
     /**
@@ -109,11 +114,9 @@ export abstract class TableQuickActionDefinitionBase {
             if (section) {
                 this.collectChildrenInSection(section, table);
             } else if (this.iconTabBar && tabKey) {
-                this.children.push({
-                    label: `'${iconTabBarfilterMap[tabKey]}' table`,
-                    enabled: true,
-                    children: []
-                });
+                const label = `'${iconTabBarfilterMap[tabKey]}' table`;
+                const child = this.createChild(label, table);
+                this.children.push(child);
                 this.tableMap[`${this.children.length - 1}`] = {
                     table,
                     iconTabBarFilterKey: tabKey,
@@ -124,7 +127,7 @@ export abstract class TableQuickActionDefinitionBase {
             }
 
             // add action id to the table map, if the service actions are needed.
-            if (this.includeServiceAction) {
+            if (this.options.includeServiceAction) {
                 const actions = await this.context.actionService.get(table.getId());
                 const actionsIds = await getActionId(table);
 
@@ -139,6 +142,30 @@ export abstract class TableQuickActionDefinitionBase {
         }
         if (this.children.length > 0) {
             this.isApplicable = true;
+        }
+    }
+
+    /**
+     * Retrieves the internal table control from a UI5Element.
+     *
+     * @param table - The UI5Element instance to analyze.
+     * @returns The internal table otherwise undefined.
+     */
+    protected getInternalTable(table: UI5Element): UI5Element | undefined {
+        try {
+            let tableInternal: ManagedObject | undefined;
+
+            if (isA<SmartTable>(SMART_TABLE_TYPE, table)) {
+                const itemsAggregation = table.getAggregation('items') as ManagedObject[];
+                tableInternal = itemsAggregation.find((item) =>
+                    [M_TABLE_TYPE, TREE_TABLE_TYPE, ANALYTICAL_TABLE_TYPE, GRID_TABLE_TYPE].some((tType) =>
+                        isA(tType, item)
+                    )
+                );
+            }
+            return tableInternal as UI5Element | undefined;
+        } catch (error) {
+            return undefined;
         }
     }
 
@@ -204,26 +231,19 @@ export abstract class TableQuickActionDefinitionBase {
             } else if (subSections.length > 1) {
                 const sectionChild = this.children.find((val) => val.label === `${section.getTitle()} section`);
                 let tableMapIndex = `${this.children.length - 1}`;
+                const label = this.getTableLabel(table);
+                const child = this.createChild(label, table);
                 if (!sectionChild) {
                     tableMapIndex = `${tableMapIndex}/0`;
+
                     this.children.push({
                         label: `'${section?.getTitle()}' section`,
                         enabled: true,
-                        children: [
-                            {
-                                label: this.getTableLabel(table),
-                                enabled: true,
-                                children: []
-                            }
-                        ]
+                        children: [child]
                     });
                 } else {
                     tableMapIndex = `${tableMapIndex}/${sectionChild.children.length - 1}`;
-                    sectionChild.children.push({
-                        label: this.getTableLabel(table),
-                        enabled: true,
-                        children: []
-                    });
+                    sectionChild.children.push(child);
                 }
 
                 this.tableMap[tableMapIndex] = {
@@ -245,11 +265,9 @@ export abstract class TableQuickActionDefinitionBase {
         sectionInfo?: { section: ObjectPageSection; subSection: ObjectPageSubSection; layout?: ObjectPageLayout }
     ): void {
         if ([SMART_TABLE_TYPE, M_TABLE_TYPE, MDC_TABLE_TYPE, TREE_TABLE_TYPE].some((type) => isA(type, table))) {
-            this.children.push({
-                label: this.getTableLabel(table),
-                enabled: true,
-                children: []
-            });
+            const label = this.getTableLabel(table);
+            const child = this.createChild(label, table);
+            this.children.push(child);
         }
 
         this.tableMap[`${this.children.length - 1}`] = {
@@ -283,5 +301,23 @@ export abstract class TableQuickActionDefinitionBase {
             title: this.context.resourceBundle.getText(this.textKey),
             children: this.children
         };
+    }
+
+    createChild(label: string, table: UI5Element): NestedQuickActionChild {
+        const child: NestedQuickActionChild = {
+            label,
+            enabled: true,
+            children: []
+        };
+        if (!this.options.areTableRowsRequired) {
+            return child;
+        }
+        const innerTable = this.getInternalTable(table);
+        const tableRows = (innerTable?.getAggregation('items') as ManagedObject[]) || [];
+        if (isA(M_TABLE_TYPE, innerTable) && !tableRows.length) {
+            child.enabled = false;
+            child.tooltip = this.context.resourceBundle.getText('TABLE_CUSTOM_COLUMN_ACTION_NOT_AVAILABLE');
+        }
+        return child;
     }
 }