+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.65
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.5.21
+
+## 0.0.64
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.5.20
+
+## 0.0.63
+
+### Patch Changes
+
+-   Updated dependencies [73475e5]
+    -   @sap-ux/ui5-config@0.26.0
+    -   @sap-ux/project-access@1.28.8
+    -   @sap-ux/system-access@0.5.19
+
 ## 0.0.62
 
 ### Patch Changes