+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.0.3
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.12
+
+## 0.0.2
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.11
+
 ## 0.0.1
 
 ### Patch Changes