+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/inquirer-common
 
+## 0.5.7
+
+### Patch Changes
+
+-   2359524: align versions of cf-tools
+-   Updated dependencies [2359524]
+    -   @sap-ux/btp-utils@0.17.1
+    -   @sap-ux/fiori-generator-shared@0.7.12
+    -   @sap-ux/telemetry@0.5.46
+
+## 0.5.6
+
+### Patch Changes
+
+-   Updated dependencies [d04a40e]
+    -   @sap-ux/feature-toggle@0.2.3
+
+## 0.5.5
+
+### Patch Changes
+
+-   @sap-ux/telemetry@0.5.45
+-   @sap-ux/fiori-generator-shared@0.7.11
+
 ## 0.5.4
 
 ### Patch Changes