+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @sap-ux/preview-middleware
 
+## 0.16.139
+
+### Patch Changes
+
+-   Updated dependencies [2359524]
+    -   @sap-ux/btp-utils@0.17.1
+    -   @sap-ux/adp-tooling@0.12.89
+
+## 0.16.138
+
+### Patch Changes
+
+-   Updated dependencies [d04a40e]
+    -   @sap-ux/feature-toggle@0.2.3
+    -   @sap-ux/adp-tooling@0.12.88
+
+## 0.16.137
+
+### Patch Changes
+
+-   0fb08df: Use ui5 version specific flp sandbox template instead of dynamic bootstrap
+
+## 0.16.136
+
+### Patch Changes
+
+-   Updated dependencies [14163cd]
+    -   @sap-ux/adp-tooling@0.12.87
+
+## 0.16.135
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.12.86
+-   @sap-ux/project-access@1.28.8
+
+## 0.16.134
+
+### Patch Changes
+
+-   c10bf9f: fix: Various lint error fixes and code improvements
+
+## 0.16.133
+
+### Patch Changes
+
+-   70e6d46: CPE - Disable Add Custom Column Quick Action, if table rows are required and not available
+
 ## 0.16.132
 
 ### Patch Changes