+++ packages/ui5-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-config
 
+## 0.26.0
+
+### Minor Changes
+
+-   73475e5: Support for multiple services and multiple annotations per service.
+
 ## 0.25.2
 
 ### Patch Changes