+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 0.0.12
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-reference-inquirer@0.3.52
+-   @sap-ux/fiori-generator-shared@0.7.12
+-   @sap-ux/telemetry@0.5.46
+
+## 0.0.11
+
+### Patch Changes
+
+-   Updated dependencies [d04a40e]
+    -   @sap-ux/feature-toggle@0.2.3
+    -   @sap-ux/ui5-library-reference-inquirer@0.3.51
+
+## 0.0.10
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.8
+-   @sap-ux/telemetry@0.5.45
+-   @sap-ux/ui5-library-reference-writer@0.1.36
+-   @sap-ux/fiori-generator-shared@0.7.11
+-   @sap-ux/ui5-library-reference-inquirer@0.3.50
+
 ## 0.0.9
 
 ### Patch Changes